rename CustomPreview.kt package to com.berlin.aflami.component
