--- presentation/designSystem/src/main/kotlin/com/berlin/aflami/component/test.kt
@@ -1,2 +1,0 @@
-package com.berlin.aflami.component
-
